in progress

+++ app/src/main/java/ru/okcode/currencyconverter/model/Config.kt
@@ -2,14 +2,14 @@ package ru.okcode.currencyconverter.model
 
 data class Config(
     val baseCurrencyCode: String,
-    val baseCurrencyAmount: Double,
+    val baseCurrencyAmount: Float,
     val visibleCurrencies: List<String>
 ) {
     companion object {
         fun getDefaultConfig(): Config {
             return Config(
                 baseCurrencyCode = "EUR",
-                baseCurrencyAmount = 1.0,
+                baseCurrencyAmount = 1f,
                 visibleCurrencies = ArrayList()
             )
         }
